First stab at async. Oh no what is happening.
Note to self: if ever writing a unit test framework, I should really get around
to writing tests for the framework itself.

That's okay, this is all unnecessary and will be re-written as Jasmine specs, I just
wanted a single supported platform to make sure the API holds up to implementation.